Fix server build errors and migration issues
- Fix migration file names in database.go (002_enrollment_tokens.sql)
- Add 006_device_management.sql to migration list
- Fix Pool() method calls across all API handlers
- Fix context type assertions in inventory_handlers.go
- Fix SendCommand signature mismatches in mobile_handlers.go
- Remove duplicate message type constants from hub.go
- Add int32 type conversion for DB connection pool config
- Add golang.org/x/oauth2 dependency

+++ server/cmd/sentinel/main.go
@@ -33,8 +33,8 @@ func main() {
 	// Initialize database with connection pool settings
 	dbConfig := &database.Config{
 		URL:      cfg.DatabaseURL,
-		MaxConns: cfg.DBMaxConns,
-		MinConns: cfg.DBMinConns,
+		MaxConns: int32(cfg.DBMaxConns),
+		MinConns: int32(cfg.DBMinConns),
 	}
 	db, err := database.NewWithConfig(dbConfig)
 	if err != nil {
